Fixed leastsquares issue with np.exp and non np.float arrays

+++ bfit/fitting/leastsquares.py
@@ -22,8 +22,8 @@ class LeastSquares:
             fn_prime_dx:    spacing in x to calculate the derivative in the case of the default calculation
         """
         self.fn = fn
-        self.x = np.asarray(x)
-        self.y = np.asarray(y)
+        self.x = np.asarray(x, dtype=np.float)
+        self.y = np.asarray(y, dtype=np.float)
         self.n = len(x)
         
         # set derivative
